update per i box

+++ Gioco/src/core/entities/Entity.java
@@ -2,14 +2,14 @@ package core.entities;
 
 public abstract class Entity {
 	protected int width, height;
-	protected int x, y;
+	protected double x, y;
 	
 	public Entity(int width, int height) {
 		this.width = width;
 		this.height = height;
 	}
 	
-	public Entity(int width, int height, int x, int y) {
+	public Entity(int width, int height, double x, double y) {
 		this.width = width;
 		this.height = height;
 		this.x = x;
@@ -36,7 +36,7 @@ public abstract class Entity {
 		this.x = x;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 	
@@ -44,7 +44,7 @@ public abstract class Entity {
 		this.y = y;
 	}
 	
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 }
